Add files via upload
Create a program that will allow a user to enter a loan amount and the number of years that they will borrow the money. Create a table of output that will display the interest rate, monthly payment, and total payment of the loan. The first interest rate will be 5% and the last will be 8%. The rate will change by .125% for each increment. Use a counter controlled loop to change the annual interest rate. Round money values to 2 decimal places.