Add files via upload
Allow a user to enter a phone number as a string that contains letters. Then convert the entire phone number into all numbers. For example:

1-800-Flowers as an input would be displayed as 1-800-356-9377

You must create a method that accepts a char and returns an int. Pass one character at a time to the method using charAt() inside a loop in main.